refactored fromPacket for execute
