+++ lib/commands/server_handshake.js
@@ -5,7 +5,6 @@ const Errors = require('../constants/errors.js');
 
 const Command = require('./command.js');
 const Packets = require('../packets/index.js');
-const Types = require('../constants/types');
 
 class ServerHandshake extends Command {
   constructor(args) {
@@ -70,6 +69,11 @@ class ServerHandshake extends Command {
     return ServerHandshake.prototype.dispatchCommands;
   }
 
+  _isStatement(query, name) {
+    const firstWord = query.split(' ')[0].toUpperCase();
+    return firstWord == name;
+  }
+
   dispatchCommands(packet, connection) {
     // command from client to server
     let knownCommand = true;
@@ -90,67 +94,7 @@ class ServerHandshake extends Command {
         break;
       case CommandCode.STMT_EXECUTE:
         if (connection.listeners('stmt_execute').length) {
-          console.log(packet)
-          const stmtId = packet.readInt32();
-          const flags = packet.readInt8();
-          const iterationCount = packet.readInt32();
-
-          let nullBitmaps = 0;
-          let i = packet.offset;
-          while (i < packet.end - 1) {
-            if((packet.buffer[i+1] == Types.VAR_STRING 
-              || packet.buffer[i+1] == Types.NULL 
-              || packet.buffer[i+1] == Types.DOUBLE
-              || packet.buffer[i+1] == Types.TINY
-              || packet.buffer[i+1] == Types.DATETIME
-              || packet.buffer[i+1] == Types.JSON) && packet.buffer[i] == 1 && packet.buffer[i+2] == 0) {
-                break
-            }
-            else {
-              nullBitmaps += packet.readInt8();
-            }
-            i++;
-          }
-
-          const types = [];
-
-          for(let i = packet.offset + 1; i < packet.end - 1; i++) {
-            if((packet.buffer[i] == Types.VAR_STRING 
-              || packet.buffer[i] == Types.NULL 
-              || packet.buffer[i] == Types.DOUBLE
-              || packet.buffer[i] == Types.TINY
-              || packet.buffer[i] == Types.DATETIME
-              || packet.buffer[i] == Types.JSON) && packet.buffer[i + 1] == 0) {
-                types.push(packet.buffer[i]);
-                packet.skip(2);
-            }
-          }
-
-          packet.skip(1);
-
-          const values = [];
-          for(let i = 0; i < types.length; i++) {
-            if(types[i] == Types.VAR_STRING) {
-              values.push(packet.readLengthCodedString(encoding))
-            }
-            else if(types[i] == Types.DOUBLE) {
-              values.push(packet.readDouble())
-            }
-            else if(types[i] == Types.TINY) {
-              values.push(packet.readInt8())
-            }
-            else if(types[i] == Types.DATETIME) {
-              values.push(packet.readDateTime())
-            }
-            else if(types[i] == Types.JSON) {
-              values.push(JSON.parse(packet.readLengthCodedString(encoding)))
-            }
-            if(types[i] == Types.NULL) {
-              values.push(null)
-            }
-          }
-          console.log(values)
-
+          const { stmtId, flags, iterationCount, values } = Packets.Execute.fromPacket(packet, encoding);
           connection.emit('stmt_execute', stmtId, flags, iterationCount, values);
         } else {
           connection.writeError({
